requirements.txt
sumatra>0.6
django-registration-redux==1.2
